Move `MaybePartial` to dedicated module
